test(endpoint): move `BenchmarkEndpoint_Type` to file `endpoint_bench_test.go`
