+++ config/endpoint/endpoint_test.go
@@ -431,20 +431,6 @@ func TestEndpoint_Type(t *testing.T) {
 	}
 }
 
-func BenchmarkEndpoint_Type(b *testing.B) {
-	for b.Loop() {
-		for _, tt := range testEndpoint_typeData {
-			endpoint := Endpoint{
-				URL:       tt.args.URL,
-				DNSConfig: tt.args.DNS,
-			}
-			if got := endpoint.Type(); got != tt.want {
-				b.Errorf("Endpoint.Type() = %v, want %v", got, tt.want)
-			}
-		}
-	}
-}
-
 func TestEndpoint_ValidateAndSetDefaults(t *testing.T) {
 	endpoint := Endpoint{
 		Name:               "website-health",
